requirements.txt
# .env support
python-dotenv>=1.0.0

# News fetching
gnews==0.2.8

# Data handling
pandas>=2.0.0

# LangChain components for RAG pipeline
langchain>=0.1.0
openai>=1.0.0
chromadb>=0.4.14

# Required for some LangChain components
tiktoken>=0.4.0

# Optional: ensures datetime parsing is fully supported
python-dateutil>=2.8.2
